Unified User data class
